break out functions into utils #16
