+++ hive/indexer/cached_post.py
@@ -1,5 +1,4 @@
 import json
-import math
 import collections
 
 from toolz import partition_all
@@ -7,8 +6,9 @@ from funcy.seqs import first
 from hive.db.methods import query, query_all, query_col, query_one
 from hive.db.db_state import DbState
 
-from hive.indexer.normalize import amount, parse_time, rep_log10, safe_img_url
-from hive.indexer.timer import Timer
+from hive.utils.post import post_stats, score
+from hive.utils.normalize import amount, parse_time, rep_log10, safe_img_url
+from hive.utils.timer import Timer
 from hive.indexer.accounts import Accounts
 from hive.indexer.steem_client import get_adapter
 
@@ -332,8 +332,8 @@ class CachedPost:
 
         # trending scores
         timestamp = parse_time(post['created']).timestamp()
-        hot_score = cls._score(rshares, timestamp, 10000)
-        trend_score = cls._score(rshares, timestamp, 480000)
+        hot_score = score(rshares, timestamp, 10000)
+        trend_score = score(rshares, timestamp, 480000)
 
         if post['body'].find('\x00') > -1:
             print("bad body: {}".format(post['body']))
@@ -343,7 +343,7 @@ class CachedPost:
         if children > 32767:
             children = 32767
 
-        stats = cls._post_stats(post)
+        stats = post_stats(post)
         body = post['body']
 
         # empty/deprecated fields
@@ -426,59 +426,9 @@ class CachedPost:
         return sqls
 
     @classmethod
-    # see: calculate_score - https://github.com/steemit/steem/blob/8cd5f688d75092298bcffaa48a543ed9b01447a6/libraries/plugins/tags/tags_plugin.cpp#L239
-    def _score(cls, rshares, created_timestamp, timescale=480000):
-        mod_score = rshares / 10000000.0
-        order = math.log10(max((abs(mod_score), 1)))
-        sign = 1 if mod_score > 0 else -1
-        return sign * order + created_timestamp / timescale
-
-    @classmethod
     def _vote_csv_row(cls, vote):
         return ','.join((vote['voter'], str(vote['rshares']), str(vote['percent']),
                          str(rep_log10(vote['reputation']))))
-
-    # see: contentStats - https://github.com/steemit/condenser/blob/master/src/app/utils/StateFunctions.js#L109
-    @classmethod
-    def _post_stats(cls, post):
-        net_rshares_adj = 0
-        neg_rshares = 0
-        total_votes = 0
-        up_votes = 0
-        for vote in post['active_votes']:
-            if vote['percent'] == 0:
-                continue
-
-            total_votes += 1
-            rshares = int(vote['rshares'])
-            sign = 1 if vote['percent'] > 0 else -1
-            if sign > 0:
-                up_votes += 1
-            if sign < 0:
-                neg_rshares += rshares
-
-            # For graying: sum rshares, but ignore neg rep users and dust downvotes
-            neg_rep = str(vote['reputation'])[0] == '-'
-            if not (neg_rep and sign < 0 and len(str(rshares)) < 11):
-                net_rshares_adj += rshares
-
-        # take negative rshares, divide by 2, truncate 10 digits (plus neg sign),
-        #   and count digits. creates a cheap log10, stake-based flag weight.
-        #   result: 1 = approx $400 of downvoting stake; 2 = $4,000; etc
-        flag_weight = max((len(str(neg_rshares / 2)) - 11, 0))
-
-        author_rep = rep_log10(post['author_reputation'])
-        is_low_value = net_rshares_adj < -9999999999
-        has_pending_payout = amount(post['pending_payout_value']) >= 0.02
-
-        return {
-            'hide': not has_pending_payout and (author_rep < 0),
-            'gray': not has_pending_payout and (author_rep < 1 or is_low_value),
-            'author_rep': author_rep,
-            'flag_weight': flag_weight,
-            'total_votes': total_votes,
-            'up_votes': up_votes
-        }
 
     @classmethod
     def _write(cls, values, mode='insert'):
@@ -487,19 +437,24 @@ class CachedPost:
     # sql builder for writing to hive_posts_cache table
     @classmethod
     def _write_sql(cls, values, mode='insert'):
+        _pk = ['post_id']
+        _table = 'hive_posts_cache'
+        assert _pk, "primary key not defined"
+        assert _table, "table not defined"
+        assert mode in ['insert', 'update'], "invalid mode %s" % mode
+
         values = collections.OrderedDict(values)
         fields = values.keys()
 
         if mode == 'insert':
             cols = ', '.join(fields)
             params = ', '.join([':'+k for k in fields])
-            sql = "INSERT INTO hive_posts_cache (%s) VALUES (%s)"
-            sql = sql % (cols, params)
+            sql = "INSERT INTO %s (%s) VALUES (%s)"
+            sql = sql % (_table, cols, params)
         elif mode == 'update':
-            update = ', '.join([k+" = :"+k for k in fields][1:])
-            sql = "UPDATE hive_posts_cache SET %s WHERE post_id = :post_id"
-            sql = sql % (update)
-        else:
-            raise Exception("unknown write mode %s" % mode)
+            update = ', '.join([k+" = :"+k for k in fields if k not in _pk])
+            where = ' AND '.join([k+" = :"+k for k in fields if k in _pk])
+            sql = "UPDATE %s SET %s WHERE %s"
+            sql = sql % (_table, update, where)
 
         return sql
