+++ hive/indexer/core.py
@@ -11,7 +11,7 @@ from toolz import partition_all
 from hive.db.methods import query
 from hive.db.db_state import DbState
 
-from hive.indexer.timer import Timer
+from hive.utils.timer import Timer
 from hive.indexer.steem_client import get_adapter
 
 from hive.indexer.blocks import Blocks
