+++ hive/utils/normalize.py
